+++ app/crud/applicant.py
@@ -9,103 +9,6 @@ from datetime import timedelta
 from fastapi import Response
 
 
-def create_applicant(db: Session, applicant_data: ApplicantCreate):
-    ## TODO programRegistered
-
-    new_applicant_id = generate_next_id(db)
-
-    new_applicant_general = ApplicantGeneralInformation(
-        applicantId=new_applicant_id,
-        nationality=applicant_data.nationality,
-        idCardNumber=applicant_data.idNumber if applicant_data.idType == "citizen" else "",
-        passportId=applicant_data.idNumber if applicant_data.idType == "passport" else "",
-        prefix=applicant_data.title,
-        firstnameTH=applicant_data.firstNameThai,
-        lastnameTH=applicant_data.lastNameThai,
-        firstnameEN=applicant_data.firstNameEnglish,
-        lastnameEN=applicant_data.lastNameEnglish,
-        submissionStatus=False,
-        password=hash_password(applicant_data.password),
-    )
-
-    new_contact = ApplicantContact(
-        applicantId=new_applicant_id,
-        applicantEmail=applicant_data.email,
-    )
-
-    applicant_models = [
-        ApplicantAddress,
-        ApplicantContactPerson,
-        ApplicantAdmissionChannel,
-        ApplicantEnglishExam,
-        ApplicantAcademicBackground,
-        ApplicantMathematicsExam,
-        ApplicantAdditionalDocuments,
-        ApplicantStatus,
-        PreliminaryEvaluation
-    ]
-
-    new_records = [model(applicantId=new_applicant_id) for model in applicant_models]
-    
-    db.add(new_applicant_general)
-    db.commit()
-
-    db.add(new_contact)
-    db.add_all(new_records)
-    db.commit()
-
-    db.refresh(new_applicant_general)
-    db.refresh(new_contact)
-    for record in new_records:
-        db.refresh(record)
-
-    return {"Message": f"Create Applicant id {new_applicant_id} Success."}
-
-
-def applicant_login(response: Response, db: Session, idNumber: str, password: str):
-    user: Optional[ApplicantGeneralInformation] = db.query(ApplicantGeneralInformation).filter(
-        (ApplicantGeneralInformation.idCardNumber == idNumber) |
-        (ApplicantGeneralInformation.passportId == idNumber)
-    ).first()
-
-    if not user or not verify_password(password, user.password):
-        raise HTTPException(status_code=401, detail="Invalid credentials")
-
-    access_token = create_token(
-        data={"sub": user.firstnameEN + " " + user.lastnameEN, "appId": user.applicantId},
-        expires_delta=timedelta(minutes=60)
-    )
-
-    refresh_token = create_token(
-        data={"sub": user.firstnameEN, "appId": user.applicantId},
-        expires_delta=timedelta(days=1),
-        secret=REFRESH_SECRET
-    )
-
-    response.set_cookie(
-        key="refresh_token",
-        value=refresh_token,
-        httponly=True,
-        secure=True,
-        samesite="strict",
-        max_age=60 * 60 * 24
-    )
-
-    return {
-        "access_token": access_token,
-        "token_type": "bearer"
-    }
-
-
-def generate_next_id(db: Session) -> str:
-    last_user = db.query(ApplicantGeneralInformation).order_by(ApplicantGeneralInformation.applicantId.desc()).first()
-    if not last_user:
-        return "0000001"
-    last_id = int(last_user.applicantId)
-    next_id = last_id + 1
-    return str(next_id).zfill(7)
-
-
 # General Information
 def update_applicant_general_info(db: Session, applicant_id: str, update_data: ApplicantGeneralInformationUpdate):
     applicant = db.query(ApplicantGeneralInformation).filter_by(applicantId=applicant_id).first()
